requirements.txt
flask
pymongo
Pillow
six
